+++ src/state.h
@@ -280,6 +280,7 @@ extern void ss_grid_init(scene_state_t *ss);
 extern void ss_grid_common_init(grid_common_t *gc);
 extern void ss_rand_init(scene_state_t *ss);
 extern void ss_midi_init(scene_state_t *ss);
+extern void ss_cal_init(scene_state_t *ss);
 
 extern void ss_set_in(scene_state_t *ss, int16_t value);
 extern void ss_set_param(scene_state_t *ss, int16_t value);