+++ tests/serialize_scene_tests.h
@@ -1,0 +1,8 @@
+#ifndef _SERIALIZE_SCENE_TESTS_H_
+#define _SERIALIZE_SCENE_TESTS_H_
+
+#include "greatest/greatest.h"
+
+SUITE_EXTERN(serialize_scene_suite);
+
+#endif